refactor: reuse GroupWidget code for NestedGroupWidget
